Adding operator< for Galaxy objects
This allows them to be compared, and therefore to be used as keys in
maps, to be put in sets, to be sorted, etc. This will be important when
moving from using GalaxyPtr to use only Galaxy objects as GalaxyPtr
objects are currently used as keys of the molgas_per_galaxy map type.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>