Avoding Subhalo copying in shark-importer
Even though we don't use this tool (yet), I realized that because we
were returning a *constant* vector of Subhalos that triggered a copy of
the value being returned. Removing the const constraint avoid the copy,
of course.

Also when Subhalos are assembled into the vector, let's create them
directly in the vector's memory instead of doing it on the stack and
then moving it into the vector.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>