Stopping usage of std::shared_ptr<Galaxy>
While using std::shared_ptr<Galaxy> was convenient as they can be easily
copied, but other than this it doesn't give us any benefit, as Galaxy
objects are always owned by a single Subhalo and referenced only
internally by Subhalos; i.e., there are no other objects holding
references galaxies contained in a Subhalo.

On top of this, two drawbacks could be also observed. The first is that
used std::shared_ptr means we need extra memory to hold the pointer
objects themselves. This amounted to a fairly small number compared to
the size of the Galaxy objects themselves, but was still unnecessary
memory usage. The second, and main, drawback was that access to Galaxy
objects always occurred through one level of memory indirection, which
would certainly not help much the cache system when looping through a
collection of galaxies, as they would not be contiguous in memory.

The change from using std::shared_ptr<Galaxy> to using Galaxy directly
in the codebase is fairly big, and considers the following aspects:

 * The main GalaxyPtr type alias is now defined as "Galaxy *" rather
than std::shared_ptr<Galaxy>

 * The Subhalo::galaxies member is now a std::vector<Galaxy> instead of
a std::vector<Galaxy *>. As a result of this change all the code
iterating over this member in range-for loops had to be ported to refer
to galaxy members via . instead of -> (e.g.; in "for (auto &galaxy:
subhalo->galaxies) { galaxy->sfr_disk = 0; }" galaxy->sfr_disk had to be
replaced by galaxy.sfr_disk).

 * The change above affected also a number of functions that previously
received GalaxyPtr objects by reference. Since we changed those to
receive Galaxy references now, we took the chance to make sure we pass
constant references when possible.

 * The central_galaxy and type1_galaxy methods still return a GalaxyPtr,
as some Subhalos are allowed to lack a central or type2 galaxy. Callers
of these methods don't need to change though, since using a "Galaxy *"
object and a std::shared_ptr<Galaxy> object is very similar.

 * To accommodate for const-correctness, a const version of the two
methods above has also been added. These return a ConstGalaxyPtr (i.e.,
"const Galaxy *").

 * The type2_galaxies method of the Subhalo class needed a bit
refurbishment, most of which was contained in an earlier commit.
Previously this method returned a newly-created vector with copies of
the GalaxyPtr objects contained in the Subhalo::galaxies member that
were of type2. Because the std::shared_ptr copies referred to the same
underlying Galaxy object they could be safely used to modify the latter,
and some code relied on this. Now that we store Galaxy objects directly
this vector-of-copies behavior doesn't satisfy this requirement anymore,
and therefore we are switching to return a type2_galaxies_view object
instead. This object allows users to iterate over the Galaxy objects
contained in the Subhalo::galaxies vector, return references to them,
but skipping those that are not type2.

 * To easily get the total number of type2 galaxies a new
type2_galaxies_count() method has been added to Subhalo.

 * The transfer_galaxies* methods in Subhalo have been reworked to avoid
copies. As a consequence the copy_galaxies_to method hhas been removed,
and the remove_galaxies() method's signature has been changed.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>