Fixed issue with stripping whih was always being applied
The option of environmental stripping was faulty and instantaneous
strangulation of the gas was always applied. This was because in the
physica_model the gas cooling function was ONLY called for centrals, and
because in galaxy mergers, the halo baryons in the satellite subhalos
were always being transferred to the central subhalo.

Now in phyical model we always call gas_cooling, and in galaxy_mergers
there is no transferring of halo baryons. This is now all done in
environment.cpp as it should have been.