Introducing the new RotatingBaryonBase class
This new class sits in between BaryonBase and Baryon and is the one
introducing the sAM member (and therefore the angular_momentum method
too).

The introduction of this class is to avoid wasting memory space in some
of the Subhalo members currently defined as Baryons (see the *_gas
members) but that actually don't seem to need an rscale defined for
them. This will save space in the size of Subhalos, and therefore on the
overall memory consumption of shark.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>