Constraining Subhalo/Galaxy members' types
A few of the members of the Subhalo and Galaxy classes were defined as
being of Baryon type unnecessarily, as they never ended up using their
sAM or rscale components. This thus represented as waste of memory
space, which could specially add up substantially in the case of
Subhalos.

The new types given in this commit are the minimum necessary for each of
these members to fully function as before. With this change the size of
the Subhalo and Galaxy classes went down from to 376/248 to 360/216
bytes, respectively.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>