Better packing for Subhalo/Galaxy classes
The order in which the members of these two classes were defined has
been quite organic until now. This has led to an inefficient packing of
data within the structure, as additional padding was being inserted in
some places due to the alignment requirements of the different data
types.

In this commit we are improving the packing of members on these two
classes by performing the two more basic optimizations we can do: bring
all those of the same type together, and putting those of bigger size
first. By doing so the size of the Subhalo class (with gcc 8) went down
from 360 to 352 bytes, while Galaxy remained the same.

Since we were rearranging the members within the class, I also took the
time to align the documentation of each member to generate better
documentation.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>