Implementing lenient galaxy equality checks
Instead of always asserting over the values read from the HDF5 file
directly for full equality, we now support first sorting them, and then
asserting that they are "close enough" (both in relative and absolute
terms). This type of equality check is useful for when, for some reason,
we change for example the order in which different structures (halos,
subhalos, galaxies) are looped over, and thus their appearance on the
output files. Changing that order can also in turn affect other values
computed by shark, because floating point is not necessarily
associative/commutative; therefore some difference can be tolerated in
the values themselves.

The default is still to check for full equality, but the new
-l/--lenient command-line flag allows for the new check variant to be
used.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>