Using std::vector to store subhalos in Halo
The only advantage of using a std::set over a std::vector to store the
Subhalo objects within a Halo was in being able to quickly check if a
Subhalo was already in the set; on the other hand using a std::set had a
big cost in memory: with gcc 8 and clang 7 it uses 48 bytes of memory,
compared to the 24 bytes of a vector. Thus, by using a vector we are
shoving off 24 bytes off the Halo class, which previously stood at 200
bytes in these two compilers.

In reality Halos have very few Subhalos within them, so using a vector
and then looking ourselves and finding out if a Subhalo was already part
of the Halo or not is not too costly, even though it was linear
complexity. On top of that we do this only once at Halo/Subhalo
construction time.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>