New range-related classes
The first class, range_filter, is a more generic form of the
recently-introduced type2_galaxies_view class. After playing with
types2_galaxies_view a bit and trying to save more memory usage in some
other parts of the code it became clear that the same idea of iterating
over a subset of a range was needed in a few places, but obviously using
different criteria. Thus, I took the types2_galaxies_view class and its
inner iterator and refactored them so they can work on generic Range and
UnaryPredicate types (and thus it's a templated class).

The second class, range, is simpler as it takes two iterators and
present them as the being and end of itself. This will make it easy to
iterate over these ranges, which can be constructed with arbitrary
iterators and not necessarily with the begin() and end() of a container,
in range-for loops, which is our users want to do most of the time.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>