Central definition of all ID types
This definition outside of the Identifiable base class allows us to use
them without having to have the full definition of the classes they are
meant to be used by, and thus compilation gets more decoupled. For
example, some exceptions expect a Halo ID or a Subhalo ID when
constructed, and until now they required the full definition of those
classes to be used.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>