Including headers more granularly
The exception.h header is heavily included by most shark modules. On the
other hand the exception.h header previously included both halo.h and
subhalo.h. These two facts meant that changing any of the two latter
triggered a re-compilation on most of the shark modules, which was
unnecessary.

Now that the exception classes use the central, not-bound-to-any-class
definition of the ID types we can stop including the halo and subhalo
headers on the exceptions header. In turn we need to explicitly include
them in the shark modules that were implicitly including them via
exceptions.h before.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>