MergerTree holding Halos in vector, not map
The indexing of HaloPtr objects by snapshot as done until now adds to
the overall memory consumption because (at least in gcc and clang) a
std::map<int, HaloPtr> is bigger than a std::vector<HaloPtr> in size.

Instead of indexing by snapshot and using more memory, we are now saving
them directly into a vector. In order to still satisfy the API needs of
MergerTree (i.e., get all halos in a given snapshot, and get the root
halo of the treee) we are using two techniques. Firstly, for efficiently
getting all halos in a snapshot we are "consolidating" a merger tree
after it has been fully populated with halos, which translates to
sorting its halos by snapshot; later when users requests halos in an
arbitrary snapshot we binary search the begin/end of the subrange on the
vector containing halos at the requested snapshot and return that as a
range object which can be looped through. This "consolidation" step is
important, as without it we would need search for halos in the entire
vector with linear complexity instead of logarithmic.

Secondly, to offer the roots of the merger tree as a range-like object
we are reusing the range_filter class with the appropriate filter
object. This still required linear complexity, but it's a much rarer
operation, and was linear before too.

This refactoring means we also don't need the static NONE member, so we
can actually remove the merger_tree.cpp file altogether.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>