Don't use temporary map to index halos by snapshot
The only advantage of this indexing was that it quickly returned the
full list of halos for a given snapshot, but on the other hand this is
one of the big small-living structures that increase the peak memory
usage of shark. Not using it takes that peak down a bit, which is
always a good thing.

The alternative now is to always iterate over all halos, and process
only those for the given snapshot. Rather than using a raw loop over all
the halos, and manually filtering halos out, we use the new range_filter
class together with the new in_snapshot predicate class. These together
then take care of "automatically" iterating over the halos for which the
condition is true, plus it gives us (with an extra iteration) the size
of the range.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>