Don't use temporary map to index halos by ID
Just like the other map we used to index halos by snapshot, the only
advantage of this one was that it quickly returned the halo for any
corresponding ID, while on the other hand this increased the peak memory
usage of shark. Not using this map removes the last peak we had in
memory usage, and is the final stepping stone in reducing the maximum
memory footprint of shark by ~20%.

The alternative now is to first sort the given halos by ID, and then use
a binary search to look for any given one. In practice this seems to be
yielding similar runtimes, but we can even tolerate some performance
degradation in favour of a lower memory footprint.

For the time being the sort_by_id and find_by_id functions are sitting
statically in the tree_builder module, but I have a local version of
these fully generalized for all shark component types, including pointer
types, which would be useful if needed in the future to avoid code
repetition.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>