Refactoring main merger tree building routine
This small refactoring of the main loop_through_halos routine is just to
manage the complexity of the function, which had become a bit difficult
to wield, with too many levels of indentation and therefore of mental
load one had to keep in order to understand what was going on.

The main change introduced in this small refactoring is two-fold:
firstly, the search for a suitable descendant subhalo for a given
halo/subhalo pair now lives in a different function. Secondly, this new
separate function is used std::find_if instead of a raw for loop for
easier code reasoning and maintenance.

Like in the previous commit, I am adding a local find_by_id function,
which I otherwise have locally available in a more generic form in case
we need to use something similar in other places of the code.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>