Fixing problem with gas size and mass in cases where SFR=0
- There was an issue where SFR=0, and the galaxy just formed a disk,
then the angular momentum calculation wouldn't update the gas disk
information because it was in an if that would only pass if the stellar
mass formed was >0. I have now separated the updating of the sAM of the
gas and the stars to avoid this issue.

- Fixed a problem with the ids of the subhalos, which were only unique
within the halo, but should have been unique for the subvolume and
snapshot.