Making PSO job monitoring more precise
Previously, after our PSO logic submitted a job to the queue we used the
job name (which we create) to check whether the job had finished or not.
This had the disadvantage of getting confused when two people were
running a PSO optimization in the same cluster, as they would end up
with similar job names at some point (we actually experienced this!).
This caused the PSO logic of a person A to wait for the jobs of the
another person B to complete, even if A's job had completed.

By using the new ability of shark-submit to report the underlying job ID
through the stdout we now catch that information and monitor that
specific job.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>