Ignore errors when removing temporary outputs
We have seen situations when shark fails to run, specially in HPC mode,
which in turn means its output base directory is not created at all,
which in turn again means we fail to delete it. We are already
"gracefully" handling runtime shark errors by assigning a very high
function value to the corresponding particles, so it doesn't make sense
to fail *after* that.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>