Multiple volumes not always to "multiple_batches"
In general, the directory structure for outputs coming from shark has
the form "<output-base>/<simulation>/<model>/<snapshot>/<subvolume>". On
the other hand, a single shark instance can process multiple subvolumes.
In these cases the <subvolume> part of the directory structure is
hardcoded to be "multiple_batches".

In the case of PSO runs with multiple subvolumes, we effectively end up
running shark instances using several subvolumes each, and therefore we
expect the "multiple_batches" directory to be present. This continues to
be true, but in preparation for "offline evaluation" of existing results
we need to take into account outputs coming not from a PSO run, but from
other executions (HPC submissions, manual runs). In those cases,
specially in the HPC case, we probably ran shark separately for each
subvolume, generating separate <subvolume> directories, which we *don't*
want to internally translate to "multiple_batches".

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>