Constraining the number of nodes to request
It has happened that users request more node than there are particles in
the swarm. This has no effect on shark, as it doesn't use MPI
communications, and therefore requesting more nodes than instances only
has the negative effect of waiting for longer in the queue, and burning
resources unnecessarily. This patch constraints the number of requested
nodes to be not greater than the swarm size.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>