Updating list of MacOS jobs
The builds based on macOS 10.12 (Sierra) have been failing due to long
installation times of gcc (ultimately needed by the hdf5 brew package),
which doesn't have a bottle for Sierra and needs to be compiled. This
compilation takes a long time, and produces a timeout on Travis, who
doesn't see output for more than 10 minutes and gives up.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>