Travis: adjusting to homebrew changes
The python3 package doesn't exist anymore, it's now simply "python", and
gets pulled anyway by cxxtest (and on top of that seems to be
pre-installed already). The python@2 package is still installed by
default though, so it needs to be uninstalled.

Also, the hdf5 package is now intalled by default, so let's not try to
install it ourselves.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>