Support for multi-threaded reproducible runs
The way we aim to support multi-threaded reproducible runs is by seeding
the RNG with a object-dependent value before using them to draw values
from the different distributions. This re-seeding, combined with the
fact that we use the ID of the particular object the operation is
related to, means the values drawn from the distributions don't depend
on the order in which we invoke the operations anymore; instead, they
now depend on the initial seed (which is potentially set by the user)
and the object being operated on.

This change means that storing the RNG as a member of the classes that
used them is no longer necessary, and it can actually be detrimental
when classes are shared by different threads, like in the case of the
DarkMatterHaloes base class. To avoid race conditions we thus simply
re-create a new RNG each time we want to draw values from one or more
distributions, seeding it with the correct value right at construction
time.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>