Avoid race conditions with random distributions
Both the lognormal and the uniform distributions used by the
DarkMatterHalos class (which is shared by multiple threads) do not have
any guarantee of being thread-safe, and indeed gcc's implementation of
the normal distribution keeps some state. This creates race conditions
between multiple threads that might be using them at the same time.

To avoid these race conditions we don't store the distributions objects
anymore as members of DarkMatterHalos, but create them from scratch each
time we need them. This shouldn't be an expensive operation, as neither
of them does calculations at construction time; but even if they did
they would probably be minor compared to the rest of the computations we
do in shark.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>