More informative messages on runtime performance
This commit adds the collection of more fine-grained information about
evolution times. This information is designed to give us a better
understanding on how parallelism is working in current and future
executions.

Firstly, we are now storing per-thread timing information, such that we
can display at the end how much time each thread spent on which task
during galaxy evolution. These numbers should be as equal as possible
across threads, which in turn would should a good spread of
computational load. We previously had this information available, but
only in a per-snapshot basis. With this new dimension we can show
aggregates on a per-snapshot or per-thread basis, plus totals for each
as well. The final report on total times is coded as a separate, public
method to allow in the future to print it from outside (e.g., a signal
handler).

Secondly, I'm adding an "evolution rate" statistic on a per-snapshot
level; that is how long it takes to evolve each galaxy. Locally I get
something in the order of tens of microseconds per galaxy, which a good
starting point for comparing different systems and executions, specially
when running with different number of threads.

Finally, a total "evolution walltime" is also calculated and shown at
the end of each shark execution. This is the final number we care about
shrinking when doing parallelism, so it's handy to have it readily
available for inspection and comparison.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>