New model of stellar tidal stripping implemented
This new model is based on the calculations of Errani et al. (2015),
which modelled the tidal stripping of satellite galaxies in a Milky-Way
halo. The outcome of the simulations were fit and here I adopt the best
fit for a cuspy core with an r_star/a=0.1. This seems to have the
desired effect on the stellar mass function.

There is a free parameter which is the minimum allowed threshold of
virial mass ratio (Mvir,now/Mvir,infall), which is relevant for galaxies
that go from centrals to becoming type II without ever becoming type 1.
If the more normal scenario of central->type I->type II takes place,
then tidal stripping is more gradual (as it depends on the virial mass
loss).