Creating more balanced partitions of merger trees
Until now the way we distribute workload across our OpenMP threads is by
simply assigning different chunks of merger trees to each thread in a
static fashion. This is a good strategy if the workload is similar
across all threads; otherwise the parallel for loop is as fast as the
slowest thread with more workload.

It turns out that odds are not necessarily in our favour, and in average
merger trees are not equally loaded with galaxies, which restricts our
parallelism. To overcome this we manually partition the list of merger
trees based on their galaxy count using a simple greedy algorithm,
producing partitions with roughly the same amount of galaxies, hence
spreading the workload better across OpenMP threads.

Initial experiments show an improvement in the walltime, with scaling
now going up to ~10x when running mini-surfs with 20 threads (it was
~<8x before). Note that galaxies live in a specific snapshot at a time,
so maybe we could study doing merger tree partitioning once per
snapshot, which should give more accurate results; this has the downside
however of being potentially more expensive in the long run, so
pros/cons would have to be weighted and compared.

Other than the partitioning itself, the other main aspect that had to be
taken care of is the fact that, without proper care, elements would have
been written out to the output files in different orders depending on
the amount of threads used by the execution. This breaks our rule of
staying reproducible and thus had to be dealt with appropriately by
sorting the halos to be written by ID to ensure the same order
regardless of how they were processed.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>