Added tidal stripping of ISM
- Added tidal stripping of the cold gas in the galaxy. Ideally we want
to solve for the radius that gives us the desired amount of stellar
stripping and use that to integrate and compute how much cold gas should
be stripped but as a quick test I'm stripping the same fraction fo colg
gas as I strip in stars (basically assuming that the ISM and stars have
the same mass distribution). The stripped gas is transferred to the hot
gas of the central subhalo.

- added a python routine to plot the stellar halo mas vs. halo mass
produced by tidal stripping.