Datasets of different size are unequal
So far we have used compare_galaxies.py automatically to determine
whether the output of two very similar shark executions are equal or
not. When the are *not* equal we have been lucky in that the number of
galaxies on each hdf5 file have been the same, but this doesn't need to
be true: the stochastic nature of the code will eventually produce
slightly different number of galaxies due to mergers happening or not
depending on other factors. This in turn means that when comparing two
hdf5 files we need to first check the shape of the datasets under the
galaxies/ group before proceeding any further, and if the datasets have
different shapes then we consider the two files not being equal.

Note that although this fixes an spurious error we suffered in our
automatic tests, it also is a good idea on its own, as the script can
now compare arbitrarily different outputs from wildly different runs and
mark them as being different.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>