Switch code to use gsl_set_error_handler_off
Until now we have been using a custom GSL error handler that raises an
exception when called. This has worked well so far, but in principle is
not guaranteed to work, as C++ exceptions do not necessarily transmit
correctly through the C layer of GSL.

This commit changes the GSL error handler used by shark to be the "off"
built-in error handler. This in turn means we need to be diligent and
check the return codes of GSL function invocations, which we do with our
new gsl_invoke function. Since our function turns error codes into
gsl_error exceptions the rest of the code doesn't need much change.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>