Make user-provided integration function GSL-safe
Just like in the case of the GSL ODE solver, until now we have always
given the GSL integration routines C++ functions for it to integrate.
These function can actually throw exception, but those exceptions might
or might not make it through the C stack of GSL (depending on compilers
and compilation flags) and therefore cannot be considered a safe way of
handling errors.

This commit creates a thin wrapper around the C++ user-provided
integration functions to catch any exceptions, turning them into GSL
error codes. This is not possible directly, as functions to be
integrated by GSL don't have a way to signal any internal errors. We
thus not only return NaN on failures, but also keep track ourselves of
errors and check for them even after a successful GSL invocation.

These changes are done in such a way that user-provided functions don't
need to care about GSL error codes, and can happily continue to throw
exceptions.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>