Use Galaxy to prime random orbit's seed values
When we introduced the thread-safe random number generation changes a
few months back we wrongly assumed that using a Halo's ID to prime the
seed number used to generate galaxy's orbit values was correct. It
actually wasn't: it has the effect that for all Galaxies in the same
Halo the same seed is generated, and therefore the same orbital values
are generated.

This commit adds a Galaxy parameter to the generate_random_orbits call
so that it can be used to prime the random number generator's seed
instead of the Halo. This should ensure results are different across
Galaxies even when they belong to the same Halo.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>