Modify halo properties to reflect output snapshot
Galaxy_writer has been modified to use the properties of the host halos
at the output snapshot at the moment of writing instead of the previous
one. This required some small changes as galaxies are transferred to
halos and subhalos after the writing happens. This should help with
matching DM distribution to galaxy positions and velocities at a given
snapshot.