Transfer galaxies before writing snapshot output
The idea of this is to make sure galaxy positions and halo properties
are consistent with the snapshot number at which the output is written
(rather than the snapshot-1).

This required several changes:
- transfer galaxies to halos of next snapshot before writing
- creation of halos_next_snapshots to loop through when writing
- new function to reset instantaneous galaxy properties (such as SFRs)
is called after the writing (rather than within transfer_galaxies)
- we now also save the birth snapshot of a galaxy to not include
galaxies in the writing that haven't been evolved yet.