Add new all_subhalos_view type
The all_subhalos_view type is a container-like class (it offers both
begin() and end() methods) that can be used to iterate over the subhalos
of a halo, as well as to count how many subhalos there are contained in
a halo. Iteration happens by visiting the central subhalo first, if any,
and then the satellites. The implementation of this visitation pattern
is implemented in the inner iterator class.

This commit introduces the new type, but doesn't update the rest of the
code to use it yet.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>