Reimplement all_subhalos using all_subhalos_view
The change was almost transparent, with some exceptions which I describe
below.

First, the code defining the central subhalo for each Halo was relying
on the first element of all_subhalos() being the most massive, which
isn't the case anymore. We do this explicitly now for this specific
case. This same code was also relying on all_subhalos() returning new
SubhaloPtr objects, so it could safely remove the satellite subhalo from
the underlying Halo::satellite_subhalos vector and still keep using it.
This isn't the case now either: the all_subhalos_view object returns
direct references to the underlying vector, so once a satellite is
removed from the vector the SubhaloPtr isn't valid anymore (most of the
times, unless someone else is pointing to it). The solution is obviously
to make a copy on this case.

The other exception was the code performing the linking between
snapshots, which removes subhalos *while iterating over them*. This
worked before because we always returned a new vector, and therefore
removing subhalos from the Halo had no effect on the iteration over the
vector. We explicitly create a new vector for this iteration now, which
then makes the code happy.

A set of unit tests also demonstrates the new functionality is working
as intended in all the possible use cases.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>