Limiting halo gas to universal baryon fraction
Testing limiting the reincorporated mass and pritine accretion to keep
the halo baryon fraction equal or below the universal baryon fraction.
This is checked as total baryon (including ejected and lost baryons), as
well as baryons inside halos (excluding ejected and lost baryons).

Another important change is the way we ensure mass growth in the trees.
The way it was done before didn't consider the many progenitors a halo
can have. I corrected that and now it's consistent.