Adding ram pressure stripping of ISM
Included the option of stripping the ISM of galaxies due to ram
pressure. This can be computed with or without the halo stripping on,
but stripping as an option needs to be switched on.

There remains one issue to consider. Currently type2 satellites don't
experience any ram pressure stripping of the ISM after becoming type 2
(only type 1 experience ram pressure stripping). This is because it is
not clear how to compute a consistent satellite galaxy position and
velocity within the halo. That's to be refined in the future.