Change in hot gas metallicity for type1 gals
This was coded in a way that the hot gas metallicity was artificially
low for type 1 satellites due to the inclusion of previously stripped
hot gas in the calculation of metallicity. What I know do, is to only
modify the gas mass that goes in the calculation of the gas density,
which is what I need to then compute a realistic cooling rate.