Added limit to accretion loop in griffin model
Added a limit in the loop for both models of accretion (warpeddisk and
selfgravitydisk), with the idea being that after a number of accretion
episodes and given the assumption of a random orientations of accretion
chunks, having a large number of accretion episodes should be
representative of what happens if we were to follow the loop until the
accretion disk exhausts its mass.

I did tests using a loop limit number from 50 to 500, and it appears
like 50 is enough for both accretion modes of the griffin20 model, and
significantly improves the performance of the code. So this is left as
the default.