Fix optim for local usage and enhance performance
The version of optim that runs shark locally was not working correctly:
it spawned multiple shark instances for each of the particles, but the
evaluation of the results was not done centrally. On top of this,
several bits and pieces of logic were misaligned with respect to the HPC
version, which as seen more use and therefore is better maintained.

This commit fixes how we run shark PSO locally. Instead of letting the
pso module invoke the optimisation function N times, we now follow a
similar strategy to the HPC case: we ask pso to issue a single call to
the optimisation function with all the particular, and within our
function we use a multiprocessing pool to run the individual shark
instances and collect the results.

Thanks to this change, the code to collect results after shark instances
have been run can be shared between the hpc and non-hpc mode. In fact,
at the top level we now always pass the same evaluation function to pso,
and then internally we switch between hpc and non-hpc only to spawn the
shark instances, with the rest of the code being common to both cases.

On top of that, I've also taken the opportunity to improve how results
are calculated and gathered. Because now we handle our own
multiprocessing pool, we are using it too to collect the results for
each particle in parallel, which makes a big difference in speed.
Additionally, the way we collected the model and error values for each
constraint evaluation was unnecessarily re-evaluating each constraint
each time, adding up to the processing times, so that has been now been
fixed.

There still remains an issue with how model and error values are stored
into a file for future usage. This has accidentally worked in the past
because there has only been a single constraint, or all the given
constrains result in module and error arrays with the same amount of
values. This isn't the case in general though, as running with
constraints that produce different amount of model/error values can
result in heterogeneous array shapes that cannot be represented as-is
with numpy, and thus don't work with the current code. To remain
backwards-compatible I've kept the code that does the dumping of these
files, but added a try/catch block around it to warn users if these
values couldn't be saved. In the future we should fix this.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>