Integrate first iteration in PSO main loop
The previous version of the code treated the first iteration somewhat
specially, but in reality this distinction is was a bit artificial, as
the first iteration isn't really that different from the rest. On the
other hand, having this split logic led to duplicated code, making it
more difficult to introduce more invasive changes.

This commit integrates the first PSO iteration into the rest of them so
that all the action happens within the main PSO loop. There are still
some code branches that are taken when running on the first iteration,
but overall this should help us bolt in the necessary changes for
supporting PSO restarts.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>