Communication iteration number down to objective function
This allows us to remove the global "count" variable, which in turn
would have made it difficult to store and load the overall PSO state.
Now that the objective function takes in an iteration number we can run
it without requiring a global state to be previously set.

I also took the chance to rename "it" to "iteration", which is easier to
understand.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>