Implement PSO state dumping/loading
This is a first implementation that is very barebones, but works: it
keeps track of the last version of the PSO internal state (best
particle/swarm positions and values, plus all particle positions and
velocities), which are dumped at the end of each iteration into a file.
When the PSO is loaded we check for the existence of this file, and if
found we then load the PSO state from there instead of creating one from
scratch. The state includes the last successful PSO iteration that was
run, which we use to start the main loop.

Note that some other state is not kept. For instance, all user-provided
options are discarded, and therefore users must still use the same
command line they used in a previous run to make sense of things.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>