Bump dependency on Boost to >= 1.68
We used to depend on >= 1.54, but in a8096298 we accidentally raised the
minimum required Boost version to 1.68 (boost::optional::has_value() was
introduced in that version to align with std::optional). This version of
Boost is 5 years old though, so instead of modifying our code to cope
with older versions, we're probably better off declaring a higher
minimum version.

This was the original cause of #13, and thus it should prevent this
issue from happening again in the near future.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>