Add spin_mass_dependence option to dark_matter_halo group
The option defaults to false, and indicates whether the lognormal
distribution used to draw lambda values should use a fix value of 0.03,
or a dynamic value based on the halo spin distribution following Kim et
al. (2015), as its m parameter (the natural logarithm of the
distribution's mean).

The Lagos et al. (2023) paper used this dependency, and therefore we're
also updating the sample configuration file that reflects that paper to
ensure reproducibility.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>