Modified AGN feedback jet mode memory model
The original model in Lagos et al. (2024, SHARK v2.0) adopted the
model of Croton et all (2016), which estipulated that the heating radius
maintained a memory and could only increase in size.

This was made more flexible by including a redshift below which
the memory if the heating radius in turned on. The idea behind is that
the high-z universe is very dynamic and halos can go on and off
hydrostatic equilibrium, while at lower redshifts this stabilises
and halos can remain on hydrostatic eq. The redshift at which
that transition happens was added as a free parameter.